+++ CarChangeApi/Domain/Dtos/CarDto.cs
@@ -4,7 +4,7 @@ namespace CarChangeApi.Domain.Dtos
 {
     public class CarDto
     {
-        public int CarId { get; set; }
+        public long CarId { get; set; }
         public CarType Type { get; set; }
         public string Mark { get; set; }
         public string Model { get; set; }
@@ -17,6 +17,6 @@ namespace CarChangeApi.Domain.Dtos
         public int Doors { get; set; }
         public GearBoxType GearboxType { get; set; }
 
-        public int AdvertisementId { get; set; }
+        public long AdvertisementId { get; set; }
     }
 }
